chore: add changeset for contracts/mcp schema move
