+++ .changeset/schema-contracts-export.md
@@ -1,0 +1,6 @@
+---
+"@outfitter/contracts": minor
+"@outfitter/mcp": patch
+---
+
+Move `zodToJsonSchema` from `@outfitter/mcp` to `@outfitter/contracts`. MCP re-exports for backward compatibility.
