+++ packages/contracts/package.json
@@ -100,6 +100,12 @@
         "default": "./dist/result/utilities.js"
       }
     },
+    "./schema": {
+      "import": {
+        "types": "./dist/schema.d.ts",
+        "default": "./dist/schema.js"
+      }
+    },
     "./serialization": {
       "import": {
         "types": "./dist/serialization.d.ts",
